chages in domain properties

+++ src/Xayah.Domain/Entities/Transaction.cs
@@ -7,12 +7,12 @@ namespace Xayah.Domain.Entities
     {
         [Key]
         public Guid Id { get; set; }
-        public int AccountId { get; set; }
-        public int BankId { get; set; }
+        public long AccountId { get; set; }
+        public long BankId { get; set; }
         public string TransactionCurrency { get; set; }
         public string TransactionType { get; set; }
         public DateTime TranscationDateTime { get; set; }
-        public int TransactionAmmount { get; set; }
+        public double TransactionAmmount { get; set; }
         public string TransactionDescription { get; set; }
     }
 }
